+++ R/ggbetweenstats.R
@@ -280,19 +280,13 @@ ggbetweenstats <- function(data,
 
   # add a logical column indicating whether a point is or is not an outlier
   data %<>%
-    dplyr::group_by(.data = ., x) %>%
-    dplyr::mutate(
-      .data = .,
-      isanoutlier = base::ifelse(
-        test = check_outlier(
-          var = y,
-          coef = outlier.coef
-        ),
-        yes = TRUE,
-        no = FALSE
-      )
-    ) %>%
-    dplyr::ungroup(x = .)
+    outlier_df(
+      data = .,
+      x = x,
+      y = y,
+      outlier.coef = outlier.coef,
+      outlier.label = outlier.label
+    )
 
   # -------------------------------- plot -----------------------------------
 
@@ -304,7 +298,7 @@ ggbetweenstats <- function(data,
     ) +
     # add all points which are not outliers
     ggplot2::geom_point(
-      data = data %>% dplyr::filter(.data = ., !isanoutlier),
+      data = dplyr::filter(.data = data, !isanoutlier),
       position = ggplot2::position_jitterdodge(
         jitter.width = point.jitter.width,
         dodge.width = point.dodge.width,
@@ -322,7 +316,7 @@ ggbetweenstats <- function(data,
     plot <- plot +
       # add all outliers in using same method
       ggplot2::geom_point(
-        data = data %>% dplyr::filter(.data = ., isanoutlier),
+        data = dplyr::filter(.data = data, isanoutlier),
         position = ggplot2::position_jitterdodge(
           jitter.width = point.jitter.width,
           dodge.width = point.dodge.width,
@@ -339,7 +333,7 @@ ggbetweenstats <- function(data,
       plot <- plot +
         # add all outliers in
         ggplot2::geom_point(
-          data = data %>% dplyr::filter(.data = ., isanoutlier),
+          data = dplyr::filter(.data = data, isanoutlier),
           size = 3,
           stroke = 0,
           alpha = 0.7,
@@ -490,25 +484,8 @@ ggbetweenstats <- function(data,
   if (isTRUE(outlier.tagging)) {
     # finding and tagging the outliers
     data_outlier_label <- data %>%
-      dplyr::group_by(.data = ., x) %>%
-      dplyr::mutate(
-        .data = .,
-        outlier = base::ifelse(
-          test = isanoutlier,
-          yes = outlier.label,
-          no = NA
-        )
-      ) %>%
-      dplyr::ungroup(x = .) %>%
       dplyr::filter(.data = ., isanoutlier) %>%
       dplyr::select(.data = ., -outlier)
-
-    # if there is no value for outlier.label
-    data_outlier_label$outlier.label <-
-      stringr::str_replace_na(
-        string = data_outlier_label$outlier.label,
-        replacement = "NA"
-      )
 
     # applying the labels to tagged outliers with ggrepel
     plot <-